clean up site map gen
